Modify expectSend code to better handle timing
The previous code worked for updating the display unit,
but failed when updating the main unit.  When the timeout
was too small, it would stop between the "Erasing" output.
If I raised it high enough to avoid that condition, it
did not recognize the "Wait FW file" at the end of erasing,
because there was not enough time between "Wait FW File" and
the 'C' that begins the request for an xmodem file.

I changed the expectSend code to use a shorter timeout and
loop looking for the expected string.